requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
firebase-admin==6.5.0
python-dotenv==1.0.1
boto3==1.34.69
Werkzeug==3.0.1 